requirements/base.txt
Django==1.8.9
psycopg2==2.6.1
bootstrap-admin==0.3.6
django-bootstrap3==7.0.0
mkdocs==0.15.3
djangorestframework